+++ src/pages/useEffect2/deps.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+import { UseEffect2Deps } from "../../demos/useEffect2/deps";
+
+const UseEffect2Page: React.FC = () => <UseEffect2Deps />;
+
+export default UseEffect2Page;